chore: move name_validation to src/account
