+++ crates/miden-protocol/src/account/name_validation.rs
