+++ crates/miden-protocol/src/account/storage/slot/mod.rs
@@ -1,5 +1,3 @@
-pub(crate) mod name_validation;
-
 mod slot_name;
 pub use slot_name::StorageSlotName;
